refactor(currency): migrate to Server Actions pattern
Server Component/Client Component パターンに移行し、Server Actions を導入しました。

主な変更:
- ページコンポーネントをServer/Client分離パターンに移行
  - page.tsx: データフェッチング用のServer Component
  - XxxContent.tsx: UI描画用のClient Component
- データミューテーション用のServer Actions実装 (actions.ts)
- エンティティベースのキャッシュ無効化戦略を追加
- 404エラーハンドリングの改善 (not-found.tsx)
- Constitution にServer Actions パターンを文書化
- Plan template にアーキテクチャガイドラインを追加
